Fix intermediate array dtype for numpy2

+++ datacube_ows/styles/component.py
@@ -163,7 +163,7 @@ class ComponentStyleDef(StyleDefBase):
                     else:
                         imgband_data = imgband_component
                 if imgband_data is None:
-                    null_np = np.zeros(list(data.dims.values()), 'uint8')
+                    null_np = np.zeros(list(data.dims.values()), 'float64')
                     imgband_data = DataArray(null_np, data.coords, data.dims.keys())
                 if imgband != "alpha":
                     imgband_data = self.compress_band(imgband, imgband_data)
